+++ app/Http/Controllers/User/ReviewController.php
@@ -32,9 +32,9 @@ class ReviewController extends Controller
     public function update(Request $request, $review_id)
     {
         $review = Models\Review::id($review_id)->first();
-        if ($review->review_status_id == 2) {
-            return response()->json(['status' => 'error', 'message' => 'Your review already in PUBLISHED'], 403);
-        }
+        // if ($review->review_status_id == 2) {
+        //     return response()->json(['status' => 'error', 'message' => 'Your review already in PUBLISHED'], 403);
+        // }
 
         $review->title = $request->title;
         $review->description = $request->description;
@@ -66,9 +66,10 @@ class ReviewController extends Controller
             }
         }
 
-        if ($review->review_status_id == 2) {
+        if ($review->review_status_id == 2 || ($review->review_status_id == 1 && $review->published_at)) {
             $this->decrementReviewData($review->id);
         }
+
         $review->delete();
         return response()->json(['status' => 'success', 'message' => 'Your Review Deleted'], 200);