+++ resources/views/components/star.blade.php
@@ -1,0 +1,2 @@
+@for ($i = 0; $i < intval($number); $i++) <i class="fa fa-star"> </i>
+  @endfor